+++ src/app/store/user/user.reducer.ts
@@ -9,16 +9,16 @@ import { createEntityAdapter, EntityAdapter, EntityState } from '@ngrx/entity';
 import { Item, items } from 'src/mockup';
 import { fetchUsers } from './user.actions';
 
-export interface Title extends EntityState<Item> {
-  selectedTitle: string | null;
+export interface User extends EntityState<Item> {
+  selectedUser: string | null;
 }
 
 export const adapter: EntityAdapter<Item> = createEntityAdapter<Item>({
   selectId: item => item.title,
 });
 
-export const initialState: Title = adapter.getInitialState({
-  selectedTitle: null,
+export const initialState: User = adapter.getInitialState({
+  selectedUser: null,
 });
 
 const _userReducer = createReducer(
@@ -31,11 +31,3 @@ const _userReducer = createReducer(
 export function userReducer(state: any, action: any) {
   return _userReducer(state, action);
 }
-
-const { selectAll, selectEntities, selectIds, selectTotal } =
-  adapter.getSelectors();
-
-export const selectUsers = createSelector(
-  (state: { counter: number; user: Title }) => state.user,
-  selectAll
-);
