+++ src/app/user/user.component.ts
@@ -3,7 +3,8 @@ import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import { Item } from 'src/mockup';
 import { fetchUsers } from '../store/user/user.actions';
-import { selectUsers, Title } from '../store/user/user.reducer';
+import { User } from '../store/user/user.reducer';
+import { selectUsers } from '../store/user/user.selector';
 
 @Component({
   selector: 'app-user',
@@ -14,12 +15,12 @@ export class UserComponent implements OnInit {
   users$: Observable<Item[]>;
   users: Item[] = [];
 
-  constructor(private store: Store<{ counter: number; user: Title }>) {
+  constructor(private store: Store<{ counter: number; user: User }>) {
     this.users$ = this.store.select(selectUsers);
 
     this.users$.subscribe(users => {
       console.log(users);
-      return (this.users = [...users]);
+      this.users = [...users];
     });
   }
